Updating MyTask, TaskCreate, TaskList
Updated MyTask by adding the necessary code to MyTask to display specific task information based on PK clicked on in TaskList.
Updated TaskList so now when you click on the task PK in the table it routes you to the MyTask view and displays the specified task information.
Updated TaskCreate to restructure how the code is displayed when adding and editing data entries is viewed.

Bugs to be fixed:
When task PK is clicked on in TaskList.vue webpage, data is not properly stored, accessed and displayed on screen in MyTask.vue webpage.